+++ src/main/java/id/kedukasi/core/serviceImpl/WilayahServiceImpl.java
@@ -58,7 +58,7 @@ public class WilayahServiceImpl implements WilayahService{
     }
 
     @Override
-    public Result getProvinsiByID(Integer provinsiId, String uri) {
+    public Result getProvinsiByID(Long provinsiId, String uri) {
         result = new Result();
         try {
             Provinsi provinsi = provinsiRepository.findById(provinsiId).get();
@@ -80,7 +80,7 @@ public class WilayahServiceImpl implements WilayahService{
     }
 
     @Override
-    public Result getAllKotaInProvinsi(Integer provinsiId, String uri) {
+    public Result getAllKotaInProvinsi(Long provinsiId, String uri) {
         result = new Result();
         try {
             Map items = new HashMap();
@@ -96,7 +96,7 @@ public class WilayahServiceImpl implements WilayahService{
     }
 
     @Override
-    public Result getKotaById(Integer kotaId, String uri) {
+    public Result getKotaById(Long kotaId, String uri) {
         result = new Result();
         try {
             Kota kota = kotaRepository.findById(kotaId).get();
@@ -118,7 +118,7 @@ public class WilayahServiceImpl implements WilayahService{
     }
 
     @Override
-    public Result getAllKecamatanInKota(Integer kotaId, String uri) {
+    public Result getAllKecamatanInKota(Long kotaId, String uri) {
         result = new Result();
         try {
             Map items = new HashMap();
@@ -134,7 +134,7 @@ public class WilayahServiceImpl implements WilayahService{
     }
 
     @Override
-    public Result getkecamatanById(Integer kecamatanId, String uri) {
+    public Result getkecamatanById(Long kecamatanId, String uri) {
         result = new Result();
         try {
             Kecamatan kecamatan = kecamatanRepository.findById(kecamatanId).get();
@@ -155,7 +155,7 @@ public class WilayahServiceImpl implements WilayahService{
     }
 
     @Override
-    public Result getAllKelurahanInKecamatan(Integer kecamatanId, String uri) {
+    public Result getAllKelurahanInKecamatan(Long kecamatanId, String uri) {
         result = new Result();
         try {
             Map items = new HashMap();
@@ -171,7 +171,7 @@ public class WilayahServiceImpl implements WilayahService{
     }
 
     @Override
-    public Result getKelurahanById(Integer kelurahanId, String uri) {
+    public Result getKelurahanById(Long kelurahanId, String uri) {
         result = new Result();
         try {
             Kelurahan kelurahan = kelurahanRepository.findById(kelurahanId).get();
